fno_trading_app/requirements.txt
# Core Dependencies
upstox-python-sdk==2.19.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# Technical Analysis
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Data Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# API & Web
requests==2.31.0
websocket-client==1.7.0

# Async Support
aiohttp==3.9.1

# Database (optional)
sqlalchemy==2.0.23

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Utilities
schedule==1.2.0
pyyaml==6.0.1
